+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [4.0.18]
+
 ### Added
 
+- enable run and debug option for Catch2 sections
 - improved `sourceFileMap` resolution
 
 ## [4.0.17] - 2021-12-27